Update boop.config.json
This is dumb but it's for a good cause! I need to see if Boop's new UI can handle multiple commands, but I don't want to make a project just for this. Don't judge me, gotta make do with that I got, and this is called a testAPI after all